[RFC] Alternative methods for root decomposition
Adds a `method` argument to `LazyTensor`'s `root_decomposition`, and implements the following methods additioal methods:
- `svd`
- `symeig`
- `pivoted_cholesky`
in addition to the existing methods (`cholesky`, `lanczos`).

These are useful when dealing with small-ish singular (or near-singular) covariances (typically posterior covariance at repeated points or points that are very close together.

This is WIP and a lot of optimizations could be done, including:
- throwing away zero eval/ singular values
- computing only non-zero evals (limited by torch)
- giving `LazyTensor` `_svd()` and `_symeig()` methods so that structured matrices can implement special cases (e.g. in case of block structure etc.)